feat: 최근 검색 키워드를 render하는 기능
