+++ src/js/handler/onModalShow.js
@@ -1,21 +1,10 @@
-import { openModal } from '../view/viewModal.js';
-
-const recentKeywordsLabel = () => {
-  return `<span class="text-gray-700">최근 검색어: </span>`;
-};
-
-const recentKeywordTemplate = (keyword) => {
-  return `<a class="chip">${keyword}</a>`;
-};
-
-const renderRecentKeywords = (recentKeywords) => {
-  $('[data-js="youtube-search-modal__recent-keywords"]').innerHTML =
-    recentKeywordsLabel() + recentKeywords.map(recentKeywordTemplate).join('');
-};
+import { openModal, renderRecentKeywords } from '../view/modal.js';
 
 export const onModalShow = () => {
   const recentKeywords = localStorage.get('recentKeywords') ?? [];
+  const saveClips = localStorage.get('savedClips') ?? [];
 
   openModal();
   renderRecentKeywords(recentKeywords);
+  renderSaveVideoCount(saveClips);
 };
