--- src/old/utils/querySelector.js
@@ -1,2 +1,0 @@
-export const $ = (target) => document.querySelector(target);
-export const $$ = (target) => document.querySelectorAll(target);
