--- src/old/view/viewModal.js
@@ -1,9 +1,0 @@
-const $modal = document.querySelector('[data-js="youtube-search-modal"]');
-
-export const openModal = () => {
-  $modal.classList.add('open');
-};
-
-export const closeModal = () => {
-  $modal.classList.remove('open');
-};
